# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Cricket Box API
  version: 1.0.0
  description: |
    ## Cricket Box - Digital Box Cricket Management Platform
    
    Complete API documentation for Cricket Box backend system.
    
    ### Features
    - User Authentication (Email/Password & OTP)
    - Room Management
    - Team Setup
    - Live Match Scoring
    - Statistics & Leaderboards
    - Admin Dashboard
    
  contact:
    name: Cricket Box Support
    email: support@cricketbox.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.cricketbox.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
...
    TossRequest:
      type: object
      required:
        - winner
        - decision
      properties:
        winner:
          type: string
          enum: [teamA, teamB]
        decision:
          type: string
          enum: [bat, bowl]
    RecordBallRequest:
      type: object
      required:
        - outcome
      properties:
        outcome:
          type: string
          enum: [dot, '1', '2', '3', '4', '6', wide, no_ball, bye, leg_bye, wicket]
        runs:
          type: integer
          minimum: 0
          maximum: 7
        isWicket:
          type: boolean
        dismissalType:
          type: string
          enum: [bowled, caught, caught_and_bowled, run_out, stumped, lbw, hit_wicket]
        batsmanOutId:
          type: string
        fielderId:
          type: string
        commentary:
          type: string
          maxLength: 200

    # Response Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errorCode:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /matches:
    get:
      tags: [Matches]
      summary: Get all matches
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: roomId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of matches
  /matches/{matchId}:
    get:
      tags: [Matches]
      summary: Get match by ID
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Match details
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /matches/{matchId}/toss:
    post:
      tags: [Matches]
      summary: Conduct toss
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TossRequest'
      responses:
        '200':
          description: Toss completed
        '400':
          description: Toss already done
  /matches/{matchId}/ball:
    post:
      tags: [Matches]
      summary: Record a ball
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordBallRequest'
      responses:
        '200':
          description: Ball recorded

    delete:
      tags: [Matches]
      summary: Undo last ball
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ball undone
        '400':
          description: No balls to undo

  # Admin Endpoints
...